Ground Check Update
Changed the ground check circle to a box for more precision. The original code is made for a player with a circular collider at the bottom, but my player uses a single box collider.

Moved comments for fall time reset to the correct position since it has been moved in an earlier commitment. Updated the descriptions too.