Wall Jump Animation Fix
Fixed the issue with jump animation not playing when jumping directly from a wall by applying an exit time.

The fall timer now correct resets when the player is on wall.

Removed the first frame of player jump animation, since it would be player in midair but it's not intended.