Animation Fix
Updated animator transitions and parameters. Jump and wall jump now uses a bool rather than a trigger.

Transition for wall jumping animation no longer has an exit time, because it was used to fix the issue with the player still being on wall for a bit after jumped, which was fixed in the previous commitment.

Adjusted wall attack animation so it works properly with the new changes.

Sit animation now only requires 1 parameter to work.