Extended Dash Invincibility
Extended invincibility time after dash to give the player a bit of grace period if their dash ends inside anything that damages them. This gives them some room for mistake.

Also gave HealthSystem script a public function to turn on invincibility for a set duration, so it can be reused at different places.