Charge Effect
Added a charge effect to visually tell the player when the charge attack is fully charged.

Currently has a bug where the timer only starts in ground state and will not reset if attack button is released outside of ground state.

The timer issue can be fixed if the two charge attack functions are called in update, but then ChargeAttack() wouldn't register in the player's walk state to transition to charge attack state.